DE_JD_compare/requirements.txt
httpx>=0.27
playwright>=1.44
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
pdfplumber>=0.11
rapidfuzz>=3.9
beautifulsoup4>=4.12
lxml>=5.2
langdetect>=1.0.9
